[media] media: move mdev list init to gobj
Let's control the topology changes inside the graph_object. So, move the
addition and removal of interfaces/entities from the mdev lists to
media_gobj_init() and media_gobj_remove().

The main reason is that mdev should have lists for all object types, as
the new MC api will require to store objects in separate places.

Change-Id: I19c66e6ca3962eded9c6e068d1827ae8bd3a6d88
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>