[media] media-entity.h: rename entity.type to entity.function
Entities should have one or more functions. Calling it as a
type proofed to not be correct, as an entity could eventually
have more than one type.

So, rename the field as function.

Please notice that this patch doesn't extend support for
multiple function entities. Such change will happen when
we have real case drivers using it.

No functional changes.

Change-Id: I30e521e988594ca132a1c42c58130732788313cc
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>