[media] dvbdev: Don't create indirect links
Indirect links are those whose the interface indirectly controls
other functions.

There are two interfaces that have indirect controls at the DVB
side:
- the network interface, with also controls the demux;
- the DVR interface with also controls the demux.

One could argue that the frontend control to the tuner is indirect.
Well, that's debateable. There's no way to create subdef interfaces
for tuner and demod, as those devices are tightly coupled. So, it
was decided that just one interface is the best to control both
entities, and there's no plan (or easy way) to decouple both. So,
the DVB frontend interface should link to both entities.

Change-Id: I6d0da412b7aabaf358c782a93080949d29688806
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>